more pep8 stuff, missed some whitespace
This is why I should clone repos even for minor changes like this... Why can't the test run something like `autopep8 --in-place --aggressive --aggressive` before commiting? (just dreaming) but that would fix problems from minor changes like this automatically and should leave the rest of the code in-tact... in theory.