Fixes issue #281
If a customer has a card on file this will remove the `Show Card` button and replace `Change Card` to `Add Card`. The hacky style fix (`style="margin-bottom:10px;"`) on line 29 should probably be moved to a class from styles.css. Also, I tested the `{% if customer.card_last_4 %}` and it works, but if there is a method `has_card`, that would obviously be more explicit and better.