Add automatic fixture updating via Stripe (#892)
Moves some of the test fixture objects to JSON files,
and adds a manage command  "./manage.py regenerate_test_fixtures"
that regenerates them from a real (test-mode) Stripe account,
and creates them in Stripe if it doesn't find them.

Then reads back the objects so we get Stripe's version of the data structure.

The command tries to reduce churn in the fixtures between runs by:

 1) faking the ids that are used for objects that we can't choose the ids
  for (most things except Plan, Product, Customer). These fake ids
  are generally stored in the object metadata for convenience
  (see FAKE_ID_METADATA_KEY).

 2) Not updating the fixture values of certain fields that we
  don't care about the value of (see model_sideeffect_fields).

Resolves #879