Change branding_icon and branding_logo to properties retrieved post-save
This commit removes the Account hack to preserve the branding_icon and branding_logo
foreign keys.
Both of them used to exist directly on the Account object but were moved to the settings
object inside of it. This meant we could no longer use them as foreign keys without
custom logic.
However, that custom logic is proving more harmful than helpful as it leads to a
chicken-and-egg problem when synchronizing the Account object.

With this change, we no longer sync them on save as FKs but instead, sync them post-save
as a simple FileUpload object without a backlink. The Account object gets a property
that retrieves the FileUpload object to preserve backwards-compatibility.