Do not auto-purge customers when deleting their associated model
This was introduced in 0eded97057f3037333443937fb7fd6cc6ed92dfb, but the behaviour is
undocumented and impossible to disable.
Concretely, this prevents implementing flows where the customer-associated model is
deleted, but the Stripe Customer is not, which may be desirable behaviour.

This is also surprising behaviour in the first place.