Mention required DJSTRIPE_FOREIGN_KEY_TO_FIELD in installation guide
I installed dj-stripe for the first time and the setup failed because DJSTRIPE_FOREIGN_KEY_TO_FIELD was missing. When looking at the settings, I discovered that this setting is actually required to operate. The docs recommend the value of `"id"` so that's what I'm listing.

(cherry picked from commit 5125e7e207bdf214d123598347104994c56a1892)