Simplified CrudType class and turned it into an Enum Class
Currently it is not possible to raise any error for non-exact matches to crudtype since we would need to have an exact mapping with all stripe events.

The current logic basically checks if an event is of crud_type, DELETE. if not it assumed it must be an UPDATE. And that should work for as sync_from_stripe_data essentially either creates or updates the concerned object as long as it can be retrieved from Stripe.