Refactored djstripe.settings to expose a djstripe_settings object
The entire application code and tests were directly importing the djstripe.settings module directly but that was causing issues with dynamically overriding settings because all-inbuilt functions modify the settings object from the django.conf module. Moreover, the quick fix taken was to reload the modules which is a bad idea in general.

The new djstripe_settings object will not only encapsulate all logic but also allow one to override settings at runtime as well. Settings code is modular and much more maintainable now.