Added ability to sync SubscriptionItem object whenever Subscription object is synced
SubscriptionItem is the only model so far that has no separate Stripe Webhook of the format SubscriptionItem.Crudtype(ish) and that is why this functinality was overlooked. All other models in the module billin have their own webhooks by Stripe, which is why this issue was not encountered in them.