Fixed Unexpected Keyword Error with autospec=True in test_payment_method
The root cause was how the method stripe.PaymentMethod.attach is implemented by Stripe. Stripe decorates the instance method attach(), which was causing the unexpected keyword argument.

The easy solution was to decorate the new classmethod stripe constructed to essentially use as a layer above this old instance method. The PR has the relevant links.