Added and/or updated __str__() of non-abstract child classes of StripeModel
Updated the __str__() of Price and Plan models to also account for Recurring subscription plans with Falt Fee and multiple Tiers as well. Origianlly, the code assumed that the billing scheme could only be "per_unit".

Updated the expires field of the Filelink model. Stripe allows it to be null but that was not the case in the code. Updated it.

Replicated the string representation of the objects on Stripe Dashboard.

In many scenarios replace str_parts() with __str__() especially if the object's string repersentation could be viewed on Stripe.

Note that it is not possible to view all objects on the dashboard and in that scenario used my judgement to show the most relevant and useful information, in my opinion.