Updated UsageRecord model and Added UsageRecordSummary model
Updated UsageRecord model with missing model fields and added the _api_create() and create() classmethods to facilitate its creation right from djstripe. In _api_create() also created and synced the UsageRecord model as it is not currently possible to retrieve it directly from Stripe.

Added UsageRecordSummary model that essentially returns summaries of Customer Usage for Metered Pricing subscriptions. It is essentially a layer on top of the UsageRecord model that Stripe exposed to return the summaries of usage in a given period. Added the api_list classmethod to allow UsageRecordSummary objects to get synced from Stripe