Updated StripeModel._stripe_object_to_record to handle DecimalField instances
This was done because the DecimalField and StripePercentField values were getting cached and the instance so returned, therefore had stale data.

Added Tests to ensure the returned instance had the updated Python object from the db and added a testmodel just for this purpose in the fields dir in the tests folder. This was done to isolate test models as recommended by Django