Added get_remote_ip function to get the originating client IP
The current implementation would have returned the IP of the client sending/forwarding the request to the Django server which would not be accurate whenever Django would be running behind a proxy which would almost be the case in production. Used the standard HTTP_X_FORWARDED_FOR header to figure out the actual IP of the originating client that initiated the request.