Updated get_default_api_key to raise an exception if api_key not found
Added a Boolean parameter, strict, to control if the method should return the platform key or raise an exception if teh apikey for the given Account instance cannot be found locally. This is important in case the user is trying to create an endpoint for a user who's api_key is not yet in the local db.