Updated APIKey.refresh_account
Wrapped self.save() in a transaction and ignore IntegrityErrors.

This was done because non-existent Platform Accounts need their owner account to get synced. And the search
for the owner account, triggers creation of the APIKey instance and this happens even before the givem method
gets "committed" and this causes the IntegrityError.