Fix APIKey Bug
* Updated APIKeyAdmin and overrode save_model()

This was done because for non-existent Platform Accounts, because of Account._find_owner_account()
it tries to retrieve by api_key, Account.get_or_retrieve_for_api_key().
Account.get_or_retrieve_for_api_key() ends up creating the APIKey as it doesn't exist at this time.

Form hasn't been saved saved let alone the associated model. This causes
an IntegrityError as the APIKey gets created before this form gets saved.

* Updated APIKey.refresh_account

Wrapped self.save() in a transaction and ignore IntegrityErrors.

This was done because non-existent Platform Accounts need their owner account to get synced. And the search
for the owner account, triggers creation of the APIKey instance and this happens even before the givem method
gets "committed" and this causes the IntegrityError.

* Fix Black error