Improve documentation about local webhook testing
I find the current documentation a bit unclear on the relationship
of DJSTRIPE_WEBHOOK_SECRET and the secret generated by Stripe CLI.

I added links to parts of Stripe's official documentation that I found
helpful.

The updated version is more verbose.
It uses the same URL as in installation docs:
/stripe/ instead of /djstripe/
And the note with duplicate information is removed.