Updated Customer.purge() method to not delete the associated idempotency_key
This was done because in the updated flow we ALWAYS generate a new
idempotency_key for a request if one is not provided. The old
flow was only applicable for Customer creation with a given user
as a subscriber. Besides re-using these keys are pointless after
a period of 24 hours anyway.